+++ lib/rfc959.c
@@ -1074,7 +1074,7 @@ rfc959_get_file (gftp_request * request, const char *filename, int fd,
   if (startsize > 0)
     {
 #if defined (_LARGEFILE_SOURCE)
-      command = g_strdup_printf ("REST %lld\r\n", startsize); 
+      command = g_strdup_printf ("REST %lld\r\n", (long long) startsize); 
 #else
       command = g_strdup_printf ("REST %ld\r\n", startsize); 
 #endif
@@ -1150,7 +1150,7 @@ rfc959_put_file (gftp_request * request, const char *filename, int fd,
   if (startsize > 0)
     {
 #if defined (_LARGEFILE_SOURCE)
-      command = g_strdup_printf ("REST %lld\r\n", startsize); 
+      command = g_strdup_printf ("REST %lld\r\n", (long long) startsize); 
 #else
       command = g_strdup_printf ("REST %ld\r\n", startsize); 
 #endif
